api_yamdb/requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
asgiref==3.8.1
atomicwrites==1.4.1
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==2.0.12
colorama==0.4.6
cryptography==44.0.0
defusedxml==0.8.0rc2
idna==3.10
iniconfig==2.0.0
oauthlib==3.2.2
packaging==24.2
pluggy==0.13.1
py==1.11.0
pycparser==2.22
python3-openid==3.2.0
pytz==2024.2
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
toml==0.10.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
